[BE] refactor: FestivalV1QueryService에서 FestivalV1QueryDslRepository을 의존하도록 변경 (#678) (#679)
* refactor: FestivalV1QueryService에서 FestivalRepository 의존을 FestivalV1QueryDslRepository 의존하도록 변경

* refactor: JPA Repository QueryDSL 의존 제거
