+++ backend/src/main/java/com/festago/festival/repository/FestivalRepositoryCustomImpl.java
@@ -1,28 +1,44 @@
 package com.festago.festival.repository;
 
+import static com.festago.festival.domain.QFestival.festival;
+
 import com.festago.festival.domain.Festival;
-import com.festago.festival.dto.FestivalV1Response;
+import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
 public class FestivalRepositoryCustomImpl implements FestivalRepositoryCustom {
 
-    private final OldFestivalRepository oldFestivalRepositoryCustom;
-
-    private final FestivalV1QueryDslRepository festivalV1QueryDslRepositoryCustom;
+    private final JPAQueryFactory queryFactory;
 
     @Override
     public List<Festival> findByFilter(FestivalFilter filter, LocalDate currentTime) {
-        return oldFestivalRepositoryCustom.findByFilter(filter, currentTime);
+        return switch (filter) {
+            case PLANNED -> plannedFestivals(currentTime);
+            case PROGRESS -> progressFestivals(currentTime);
+            case END -> endFestivals(currentTime);
+        };
     }
 
-    @Override
-    public Slice<FestivalV1Response> findBy(FestivalSearchCondition searchCondition) {
-        return festivalV1QueryDslRepositoryCustom.findBy(searchCondition);
+    private List<Festival> plannedFestivals(LocalDate currentTime) {
+        return queryFactory.selectFrom(festival)
+            .where(festival.startDate.gt(currentTime))
+            .orderBy(festival.startDate.asc()).fetch();
+    }
+
+    private List<Festival> progressFestivals(LocalDate currentTime) {
+        return queryFactory.selectFrom(festival)
+            .where(festival.startDate.loe(currentTime).and(festival.endDate.goe(currentTime)))
+            .orderBy(festival.startDate.asc()).fetch();
+    }
+
+    private List<Festival> endFestivals(LocalDate currentTime) {
+        return queryFactory.selectFrom(festival)
+            .where(festival.endDate.lt(currentTime))
+            .orderBy(festival.endDate.desc()).fetch();
     }
 }
