+++ backend/src/test/java/com/festago/support/RepositoryTest.java
@@ -12,7 +12,7 @@ import org.springframework.context.annotation.Import;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @DataJpaTest(showSql = false)
-@Import({QuerydslConfig.class, JpaAuditingConfig.class, RepositoryBeanConfig.class})
+@Import({QuerydslConfig.class, JpaAuditingConfig.class})
 public @interface RepositoryTest {
 
 }
